Rename logging module to avoid conflicts.
